Adding templating to matrix and cleaning up function arguments + docs

+++ includes/src_includes/interpolation.h
@@ -3,4 +3,4 @@
 #include "matrix.h"
 
 
-void interpolate_vector(const std::vector<double> &coarse_array, std::vector<double> &fine_array);
+void interpolate_vector(const std::vector<float> &coarse_array, std::vector<float> &fine_array);
